Move `nix flake {new,init}` at toplevel
Fix #4502
